# vent/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Vent Backend
host: 'localhost:5000'
basePath: /
...
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/CreateCompetition:
     post:
       tags:
         - competition
       summary: Create Competition
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: title
          in: query
          description: The title of competition
          required: true
          type: string
        - name: standings
          in: query
          description: The ObjectId for a standings
          required: true
          type: object
          format: ObjectId
        - name: fixtures
          in: query
          description: The ObjectId for a fixtures
          required: true
          type: object
          format: ObjectId
        - name: archive
          in: query
          description: The ObjectId for a archive
          required: true
          type: object
          format: ObjectId
        - name: results
          in: query
          description: The ObjectId for a results
          required: true
          type: object
          format: ObjectId
       responses:
         default:
           description: successful operation
  /api/AddArchive:
     post:
       tags:
         - competition
       summary: Create Archive and adding it to db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: url
          in: body
          description: Create Archive
          required: true
          type: string
       responses:
         default:
           description: successful operation
  /api/AddResults:
     post:
       tags:
         - competition
       summary: Create Results table and adding it to db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: url
          in: body
          description: Create Results
          required: true
          schema:
       responses:
         default:
           description: successful operation
  /api/AddStandings:
     post:
       tags:
         - competition
       summary: Create Standings table and adding it to db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: url
          in: body
          description: Create Standings
          required: true
       responses:
         default:
           description: successful operation
  /api/DeleteCompetition:
     post:
       tags:
         - competition
       summary: Delete competition from db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of competition to be deleted
          required: true
          schema:
       responses:
         default:
           description: successful operation
  /api/DeleteFixtures:
     post:
       tags:
         - competition
       summary: Delete fixtures from db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of fixtures to be deleted
          required: true
          schema:
       responses:
         default:
           description: successful operation
  /api/DeleteResults:
     post:
       tags:
         - competition
       summary: Delete results from db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of results to be deleted
          required: true
          schema:
       responses:
         default:
           description: successful operation
  /api/DeleteStandings:
     post:
       tags:
         - competition
       summary: Delete standings from db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of satndings to be deleted
          required: true
          schema:
       responses:
         default:
           description: successful operation
  /api/DeleteArchive:
     post:
       tags:
         - competition
       summary: Delete archive from db
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of archive to be deleted
          required: true
          schema:
       responses:
         default:
           description: successful operation        
  /api/UpdateFixtures:
     post:
       tags:
         - competition
       summary: update competition's fixtures 
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: fixid 
          in: body
          description: id of new fixtures 
          required: true
        - name: competid
          in: body
          description: id of competition to be updated
          required: true
       responses:
         default:
           description: successful operation
  /api/UpdateResults:
     post:
       tags:
         - competition
       summary: Update results of a competition 
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: resid
          in: body
          description: id of results 
          required: true
        - name: resid
          in: body
          description: id of competition to be updated
          required: true
       responses:
         default:
           description: successful operation
  /api/UpdateStandings:
     post:
       tags:
         - competition
       summary: Update a competition's standings 
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: id
          in: body
          description: id of satndings 
          required: true
        - name: id
          in: body
          description: id of competition to be updated 
          required: true
       responses:
         default:
           description: successful operation
  /api/UpdateArchive:
     post:
       tags:
         - competition
       summary: update competition's archive
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: archid
          in: body
          description: id of archive
          required: true
        - name: competid
          in: body
          description: id of competition to be updated
          required: true
       responses:
         default:
           description: successful operation        
...